+++ apps/api/controllers/atleisure/booking.rb
@@ -13,5 +13,8 @@ module API::Controllers::AtLeisure
       AtLeisure::Client.new(credentials).book(params)
     end
 
+    def supplier_name
+      AtLeisure::Client::SUPPLIER_NAME
+    end
   end
 end