+++ CHANGELOG.md
@@ -11,10 +11,13 @@ how this file is formatted and how the process works.
 
 ### Removed
 - removed `fee` field from `Quotation`. Fees should be included in the total.
+- removed `errors` field from `Quotation` and `Reservation`
+- removed `successful?` method from `Quotation` and `Reservation`
 
 ### Changed
 - moved `JSONRPC` to `lib` folder
 - customized timeout option for `HTTPClient`
+- all clients return `Result` for both `quote` and `book` methods
 
 ## [0.4.3] - 2016-06-29
 ### Changed